Address review: narrower exceptions, CLI aliases, shared redaction
